+++ server/TWS Foundation/TWS Foundation/Controllers/Security/AccountsController.cs
@@ -32,4 +32,9 @@ public class AccountsController
     public async Task<IActionResult> Update(Account Account) {
         return Ok(await Service.Update(Account));
     }
+
+    [HttpPost(), Auth("", "")]
+    public async Task<IActionResult> GetPermits(Account Account) {
+        return Ok(await Service.GetPermits(Account));
+    }
 }