+++ server/TWS Foundation/CHANGELOG.md
@@ -32,7 +32,8 @@
     17. Added [Address] view service.
     18. Added [Business] permits in service controllers.
     19. Added [Contact] create and update controllers.
-    20. Added the folowing [TWSSecurity] depots:
+    20. Added [Accounts] [GetPermits] controller.
+    21. Added the folowing [TWSSecurity] depots:
         - [AccountsPermits]
         - [AccountsProfiles]
         - [Permits]